+++ clioinfra/unpivot_pop.sql
@@ -5,7 +5,7 @@ create extension if not exists hstore;
 
 CREATE table geoinfra.pop_unpivot
 AS 
-SELECT year, (h).key as name, (h).value As amount
+SELECT year, (h).key as name, (h).value::float As amount
  FROM (SELECT year, each(hstore(foo) - 'year'::text) As h
   FROM geoinfra.pop as foo  ) As unpiv ;
